Shift renameAttrib and renameSlot to bootstrapVueProcessor
